# libobs: Separate textures for YUV input