… libobs/obs-source.c