### Description
The shaders to unpack YUV information from the same texture were rather
complicated. Breaking them up into separate textures makes the shaders
much simpler, and we can remove the PRECISION_OFFSET hack.

Performance also gets a nice boost on Intel for planar textures.

### Motivation and Context
Looking for more GPU headroom, and it's there. Also nice to remove shader hacks.

### How Has This Been Tested?
- [x] Verify all modified code paths work via breakpoint inspection
- [x] Deinterlacting on/off
- [x] Sanity check 1 input + 1 output on Mac
- [x] Use custom plugin to check all input formats on both D3D and GL
	UYVY
	YUY2
	YVYU
	I420
	I422
	I444
	NV12
	Y800 (limited/full)
	RGB (limited/full)
	BGR3 (limited/full)
- [x] Retest output formats on both D3D and GL: i420, i444, NV12, RGB

Intel GPA, SetStablePowerState, Intel HD Graphics 530, 1920x1080

UYVY: 473 us -> 457 us
YUY2: 492 us -> 422 us
YVYU: 491 us -> 441 us
I420: 1637 us -> 505 us
I422: 1644 us -> 482 us
I444: 1653 us -> 504 us
NV12: 1656 us -> 369 us
Y800 (limited): 270 us -> 277 us
Y800 (full): 263 us -> 289 us
RGB (limited): 341 us -> 411 us
BGR3 (limited): 512 us -> 509 us
BGR3 (full): 527 us -> 534 us

### Types of changes
- Performance enhancement (non-breaking change which improves efficiency)
- Code cleanup (non-breaking change which makes code smaller or more readable)

### Checklist:
- [x] My code has been run through [clang-format](https://github.com/obsproject/obs-studio/blob/master/.clang-format).
- [x] I have read the [**contributing** document](https://github.com/obsproject/obs-studio/blob/master/CONTRIBUTING.rst).
- [x] My code is not on the master branch.
- [x] The code has been tested.
- [x] All commit messages are properly formatted and commits squashed where appropriate.